Test and instructions improvements
