+++ catapult_os_template.yaml
@@ -52,10 +52,10 @@ objects:
             value: ${KONTINUITY_CATAPULT_GITHUB_APP_CLIENT_ID}
           - name: KONTINUITY_CATAPULT_GITHUB_APP_CLIENT_SECRET
             value: ${KONTINUITY_CATAPULT_GITHUB_APP_CLIENT_SECRET}
-          - name: KEYCLOAK_URL
-            value: ${KEYCLOAK_URL}
-          - name: KEYCLOAK_REALM
-            value: ${KEYCLOAK_REALM}
+          - name: CATAPULT_KEYCLOAK_URL
+            value: ${CATAPULT_KEYCLOAK_URL}
+          - name: CATAPULT_KEYCLOAK_REALM
+            value: ${CATAPULT_KEYCLOAK_REALM}
           image: ""
           imagePullPolicy: Always
           name: ${NAME}
@@ -137,9 +137,9 @@ parameters:
   value: https://10.1.2.2:8443
 - description: The URL (without the /auth part) of a Keycloak installation to perform SSO authentication
   displayName: KeyCloak URL
-  name: KEYCLOAK_URL
+  name: CATAPULT_KEYCLOAK_URL
   value: http://sso.prod-preview.openshift.io
 - description: The keycloak realm to be used
   displayName: KeyCloak Realm
-  name: KEYCLOAK_REALM
+  name: CATAPULT_KEYCLOAK_REALM
   value: fabric8
